requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
python-dotenv==1.0.0
notion-client==2.1.0
watchdog==3.0.0
python-multipart==0.0.6
pyyaml==6.0.1
pytest==8.4.2
pytest-asyncio==0.23.2
requests==2.31.0
# celery==5.3.6 # Optional, for background tasks
# redis==5.0.1 # Optional, if using Celery
# openai==1.3.7 # Optional, for external AI services
# anthropic==0.7.0 # Optional, for external AI services